+++ starsky/starsky/clientapp/src/components/organisms/preferences-password/preferences-password.stories.tsx
@@ -1,6 +1,6 @@
 import { storiesOf } from "@storybook/react";
 import React from "react";
-import { PreferencesPassword } from './preferences-password';
+import PreferencesPassword from './preferences-password';
 
 storiesOf("components/organisms/preferences-password", module)
   .add("default", () => {